Migrate to the new types
Start using the Stacks, StackDetails other recursive type dependencies in the Desktop application.

- Adapt the code to fit the expected (correct) type information

+++ apps/desktop/cypress/e2e/support/scenarios/unappliedBranchesAndTargetCommits.ts
@@ -38,7 +38,7 @@ export default class UnappliedBranchesAndTargetCommits extends MockBackend {
 			createMockCommit({
 				id: `target-commit-${index}`,
 				message,
-				createdAt: Date.now() - index * 1000
+				createdAt: BigInt(Date.now() - index * 1000)
 			})
 		);
 	}
